requirements.txt
sounddevice
numpy
openai-whisper
pyautogui
keyboard
pystray
Pillow 